move unit tests out of v2
