+++ test/support/app.js
@@ -1,7 +1,7 @@
 'use strict'
 
 // hacky way to stop the app logging
-const config = require('../../../config')
+const config = require('../../config')
 config.loggly = false
 
 // gag ssl-root-cas
@@ -18,7 +18,7 @@ exports.init = async function () {
 
   // start the app if not already
   if (!app) {
-    app = require('../../../index')
+    app = require('../../index')
     await new Promise((resolve, reject) => {
       app.on('listening', (err) => {
         if (err) {
